UPSTREAM: mt76: fix dfs state issue with 160 MHz channels
When operating on a mix of DFS and non-DFS channels, the driver only checks
the CAC status of the control channel. This causes beacons/tx to fail if the
control channel is on a non-DFS channel.
Fix this by calling cfg80211_reg_can_beacon to determine the DFS status of
all affected channels

Signed-off-by: Felix Fietkau <nbd@nbd.name>
(cherry picked from commit 00a883e6806e394638d344ba281c2727d6ffcbd4
 https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

BUG=b:178754244
TEST=Build on corsola

Signed-off-by: Sean Wang <objelf@gmail.com>
Change-Id: I0e22a785af0d00f76187bd72899eaa158674a039