mt76: mt7921: fix race issue between reset and suspend/resume
It is unexpected that the reset work is running simultaneously with
suspend/resume context.

The reset work should be synchronized at the beginning of the suspend
and subsequent reset requests should be ignored until at the end of
suspend/resume handler to ensure that there is no reset task running
simultaneously.

In case there is an error that happens during either suspend or resume
handler, we would break immediately and explicitly schedule a reset
task to recover the error before return the error code.

Co-developed-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: YN Chen <YN.Chen@mediatek.com>
Change-Id: I6e5867277eeb2edb8081d8401965ee97fc72cde6