WIP: FROMLIST: mt76: sdio: poll sta stat when device transmits data
It is not meaningful to poll sta stat when there is no the data traffic.
So polling sta stat when device have transmitted data instead to save
CPU power.

That implies that it is unallowed the stat_work to work while MCU is
being initialized in the really early stage. That is a required patch
for ("mt76: mt7663s: fix the deadlock caused by sdio->stat_work")
because mt7615_mcu_set_drv_ctrl pointer isn't set done until MCU is
ready.

BUG=b:231903123
TEST=emerge-corsola chromeos-kernel-5_15

Fixes: d39b52e31aa6 ("mt76: introduce mt76_sdio module")
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Change-Id: I6508e918b33f37e8ed9a7c64bb7a79e1724e2cf0