mt76: mt7921: get rid of false postive reset
False postive reset would be possibly triggered by those cmds we applied
in suspend with HZ mcu timeout, especially it happened when we enabled
kernel log in pm core to diagnose how much time we spend in each driver
during suspend procedure. So we enlarge the value and align the mcu timeout
as others cmds we did to reduce the false postive reset.

Signed-off-by: Sean Wang <sean.wang@mediatek.com>