wifi: mt76: mt7921: drop redudant the rx cleanup work
mt76_dma_rx_cleanup would be frequenetly called up to reset the dma rings
to be freshed as new ones when switching back from the deep sleep mode to
the active mode on mt7921 and mt7922.

We have already did in before the chip reset or wfdma reset. Thus,
we don't have to do it again to further improve the scheduling latency
during the deep sleep mode.

Signed-off-by: Sean Wang <sean.wang@mediatek.com>